+++ src/main/java/com/flytrap/rssreader/infrastructure/api/alert/AlertSender.java
@@ -1,9 +1,9 @@
-package com.flytrap.rssreader.service.alert.platform;
+package com.flytrap.rssreader.infrastructure.api.alert;
 
 import com.flytrap.rssreader.domain.alert.AlertPlatform;
 import com.flytrap.rssreader.service.dto.AlertParam;
 
-public interface AlarmService {
+public interface AlertSender {
 
     boolean isSupport(AlertPlatform alertPlatform);
     void sendAlert(AlertParam value);
